pyproject.toml
[tool.poetry]
name = "auditless-py-solc-x"
version = "0.1.0"
description = ""
authors = []

[tool.poetry.dependencies]
python = "^3.8"
py-solc-x = "^1.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "*"
mypy = "*"
isort = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
